+++ reqherd/webservice/models/harddes.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Unicode, BigInteger, ForeignKey
+from sqlalchemy import Column, Unicode, Integer, ForeignKey
 
 from ..db.baseclass import Requirement
 
@@ -7,7 +7,7 @@ class Hardware_Design_Element(Requirement):
 
     doc_prefix = Column(Unicode(10), nullable=False, default="HWD")
     hardware_requirement_id = Column(
-        BigInteger, ForeignKey("hardware_requirement.id"), nullable=False
+        Integer, ForeignKey("hardware_requirement.id"), nullable=False
     )
 
     def __repr__(self):
