+++ reqherd/webservice/models/hardreq.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Unicode, BigInteger, ForeignKey
+from sqlalchemy import Column, Unicode, Integer, ForeignKey
 
 from ..db.baseclass import Requirement
 
@@ -7,7 +7,7 @@ class Hardware_Requirement(Requirement):
 
     doc_prefix = Column(Unicode(10), nullable=False, default="HWRS")
     system_requirement_id = Column(
-        BigInteger, ForeignKey("system_requirement.id"), nullable=False
+        Integer, ForeignKey("system_requirement.id"), nullable=False
     )
 
     def __repr__(self):
